+++ src/editor/adapters/attempting-editor.ts
@@ -7,7 +7,7 @@ import {
   Modification,
   RelativePath
 } from "../editor";
-import { Source } from "../highlights";
+import { Source } from "../../highlights/highlights";
 import { Position } from "../position";
 import { Selection } from "../selection";
