+++ src/editor/editor.ts
@@ -2,7 +2,7 @@ import { Selection } from "./selection";
 import { RelativePath } from "./path";
 import { Position } from "./position";
 import { ErrorReason } from "./error-reason";
-import { Source } from "./highlights";
+import { Source } from "../highlights/highlights";
 
 export { AbsolutePath, RelativePath } from "./path";
 export { ErrorReason, toString as errorReasonToString } from "./error-reason";
